Deployment/Deploy/requirements.txt
Flask
numpy
Pillow
keras==3.8.0
tensorflow==2.18.0
werkzeug
gunicorn
